API users get-by-id and get-by-email paths
